✨ feat: refactor sidebar user component and integrate reminder preferences dialog; remove deprecated settings page and related components
