+++ src/data/sidebar-data.ts
@@ -1,12 +1,10 @@
 import {
   CalendarDays,
-  CreditCard,
   Globe,
   Home,
   LifeBuoy,
   PenTool,
   Send,
-  Settings,
 } from "lucide-react";
 
 export const sidebarData = {
@@ -15,8 +13,6 @@ export const sidebarData = {
     { title: "Explore", url: "/explore", icon: Globe },
     { title: "Itineraries", url: "/itineraries", icon: CalendarDays },
     { title: "Blogs", url: "/blogs", icon: PenTool },
-    { title: "Credits", url: "/credits", icon: CreditCard },
-    { title: "Settings", url: "/settings", icon: Settings },
   ],
   sidebarSecondary: [
     { title: "Support", url: "/support", icon: LifeBuoy },
